third_party/WebKit/LayoutTests/imported/web-platform-tests/webrtc/simplecall-expected.txt

CONSOLE ERROR: line 38: Uncaught ReferenceError: RTCPeerConnection is not defined
This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught ReferenceError: RTCPeerConnection is not defined
Harness: the test ran to completion.
